+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added a `SqliteDocumentStore` that uses a SQLite database to store documents.
 - Added a `PostgresDocumentStore` that uses a PostgreSQL database to store documents.
+- Added a `TxtDocumentStore` that reads documents from a single text file, separated by
+  newlines.
 
 ### Changed
 - Added defaults to all arguments in each component's constructor, so that the